+++ Order2/src/main/java/org/example/DTO/PurchaseDTO.java
@@ -5,8 +5,8 @@ import jakarta.validation.constraints.NotNull;
 
 public record PurchaseDTO(
         @NotNull(message = "productId cannot be null")
-        Integer productId,
+        Long productId,
         @Min(value = 0, message = "quantity cannot be less than 0")
-        Integer quantity
+        Long quantity
 ) {
 }
