+++ Order2/src/main/java/org/example/Exception/OrderItemNotFoundException.java
@@ -1,7 +1,7 @@
 package org.example.Exception;
 
 public class OrderItemNotFoundException extends RuntimeException {
-    public OrderItemNotFoundException(int id) {
+    public OrderItemNotFoundException(long id) {
         super("Order Item with id: "+ id +" is not found");
     }
 }
