+++ Order2/src/main/java/org/example/Exception/OrderNotFoundException.java
@@ -1,7 +1,7 @@
 package org.example.Exception;
 
 public class OrderNotFoundException extends RuntimeException {
-    public OrderNotFoundException(int orderId) {
+    public OrderNotFoundException(long orderId) {
         super("Order with order id: "+ orderId+" is not found");
     }
 }
