+++ Order2/src/main/java/org/example/Repository/OrderItemDao.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderItemDao extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemDao extends JpaRepository<OrderItem, Long> {
 
 //    @Query("SELECT oi FROM OrderItem oi WHERE oi.orderId = :orderId and oi.productId = :productId")
 //    OrderItem findByOrderIdAndProductId(@Param("orderId") Long orderId, @Param("productId") Long productId);
 
-    List<OrderItem> findByOrderId(Integer orderId);
+    List<OrderItem> findByOrderId(Long orderId);
 }
